Update 1.07
- The SmartWordlist method is now multithreaded for increased
performance on multi-core processors.
- The SmartWordlist method now automatically saves its progress every 30
seconds. If the tool is interrupted before it finishes checking all
combinations in the word list it will resume from the last save.
- Added new -firstchars parameter for use with SmartWordlist method.
This parameter allows the user to specify known characters at the start
of the password. For example, if the user knows that the forgotten
password starts with "Apple", the tool can insert "Apple" at the front
of tested passwords.
- Added new -pieces parameter for use with SmartWordlist method. This
parameter allows the user to specify minimum and maximum number of word
pieces to use when building passwords. The tool will start at the first
combination of minimum word pieces and stop at the last combination of
maximum word pieces.